requirements.txt
# Core GIS Libraries
geopandas>=0.14.0
rasterio>=1.3.0
folium>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web and API
requests>=2.31.0
httpx>=0.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# EPA AQI API specific dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0
